+ rename drivers to outputs following PR request
