+++ curiefense/curielogger/pkg/logger.go
@@ -9,26 +9,26 @@ import (
 )
 
 type LogSender struct {
-	exportDrivers io.WriteCloser
-	metrics       *Metrics
+	output  io.WriteCloser
+	metrics *Metrics
 
 	closed *atomic.Bool
 }
 
-func NewLogSender(v *viper.Viper, drivers io.WriteCloser, metrics *Metrics) *LogSender {
-	return &LogSender{exportDrivers: drivers, metrics: metrics, closed: atomic.NewBool(false)}
+func NewLogSender(v *viper.Viper, output io.WriteCloser, metrics *Metrics) *LogSender {
+	return &LogSender{output: output, metrics: metrics, closed: atomic.NewBool(false)}
 }
 
 func (ls *LogSender) Write(l *entities.LogEntry) error {
 	ls.metrics.add(l)
 	b, _ := jsoniter.Marshal(l.CfLog)
-	_, err := ls.exportDrivers.Write(b)
+	_, err := ls.output.Write(b)
 	return err
 }
 
 func (ls *LogSender) Close() error {
 	ls.closed.Store(true)
-	return ls.exportDrivers.Close()
+	return ls.output.Close()
 }
 
 func (ls *LogSender) Closed() bool {
