+++ curiefense/curielogger/pkg/outputs/buckets.go
@@ -1,4 +1,4 @@
-package drivers
+package outputs
 
 import (
 	"compress/gzip"
