+++ curiefense/curielogger/pkg/outputs/fluentd.go
@@ -1,4 +1,4 @@
-package drivers
+package outputs
 
 import (
 	"fmt"
